use the same route definitions in eskip and routing
