+++ routing/pathgen_test.go
@@ -1,4 +1,4 @@
-package requestmatch
+package routing
 
 import (
 	"math/rand"
